+++ src/utils/constants.ts
@@ -1,6 +1,6 @@
 import { LABEL_POSITION, TELESCOPE_LOW, TELESCOPE_MID } from '@ska-telescope/ska-gui-components';
-import { env } from '../env';
 import Target from './types/target';
+import { env } from '@/env';
 
 export const USE_LOCAL_DATA = env.REACT_APP_USE_LOCAL_DATA === 'true';
 export const USE_LOCAL_DATA_SENSITIVITY_CALC =
@@ -22,8 +22,10 @@ export const MSENTRA_CLIENT_ID = env.MSENTRA_CLIENT_ID;
 export const MSENTRA_TENANT_ID = env.MSENTRA_TENANT_ID;
 export const MSENTRA_REDIRECT_URI = env.MSENTRA_REDIRECT_URI;
 export const MSENTRA_API_URI = 'api://e4d6bb9b-cdd0-46c4-b30a-d045091b501b';
-
-export const APP_OVERRIDE_GROUPS = env.REACT_APP_OVERRIDE_GROUPS; // Set to Let for testing purposes.
+//
+export const APP_OVERRIDE_GROUPS = window.Cypress
+  ? localStorage.getItem('cypress:group')
+  : env.REACT_APP_OVERRIDE_GROUPS;
 
 /*****************************************/